Fix invalid year in file/directory info.

+++ OS/Kernel/FileSystems/Generic/filesystem_time.h
@@ -5,7 +5,7 @@
 
 typedef struct filesystem_time
 {
-    uint8_t year;
+    uint16_t year;
     uint8_t month;
     uint8_t day;
     uint8_t hour;
